Add metadata_only flag to GET /v1/credentials/<ID> (#274)
This commit does 2 things.

Add a metadata_only flag to GET /v1/credentials/. This provides more granular controls for fetching a credential. One benefit this has is that it's very easy to tell when a credential_pair is sent in the response (when metadata_only=False). This gives us an easier way to see when a credential was last decrypted

Update the credential's last decrypted time if the credential_pairs was sent in the response. NOTE: this is opt in - there is a env variable that must be set for the save to occur. We make this opt in to not disturb existing users