Add typing to confidant unit tests (#392)
Typing was added to confidant unit tests. This is to further improve the
service.

`run-piptools` was utilized for adding the new requirements and the mypy
configurations were added per the infradocs.

Some versions were pinned due to conflicting deprecation of functions,
mainly flask (and the flask dependencies) due to a change in
[werkzeug](https://werkzeug.palletsprojects.com/en/2.1.x/changes/#version-2-1-0)
which deprecated the `safe_str_cmp` function.

These changes were tested locally by running the service per the
[docs](https://lyft.github.io/confidant/contributing.html).
`mypy` was added to the `pre-commit` configuration. `pre-commit` was
utilized as another avenue for testing everything is working correctly.
Finally `pytest` was ran to ensure all the unit tests were passing
correctly.